Split files and add net rating and fourfactor
